+++ api/db/migrate/20201204083527_create_jourferiers.rb
@@ -4,7 +4,7 @@ class CreateJourferiers < ActiveRecord::Migration[6.0]
       t.string :evenement
       t.date :dateferie
       t.integer :anne
-      t.integer :surplus
+      t.float :surplus
 
       t.timestamps
     end
